% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_w.R
\name{calculate_w}
\alias{calculate_w}
\title{Calculate FFF channel thickness}
\usage{
calculate_w(
  t1,
  D,
  focus = 10,
  transition = 1,
  w = 500/1e+06,
  Vc = 0.0025,
  Vout = 0.001,
  Vin = 5e-04,
  temp = 25,
  eta = 0.00089,
  dims = fffprocessr:::chamber_dims,
  tol = 1e-06,
  maxiter = 10
)
}
\arguments{
\item{t1}{Retention time, in minutes.}

\item{D}{Diffusion coefficient of standard.}

\item{focus}{Focus period, in minutes.}

\item{transition}{Transition time between focusing and elution, in minutes.}

\item{w}{Nominal channel thickness, in metres.}

\item{Vc}{Cross-flow rate, in L/min.}

\item{Vout}{Detector flow rate, in L/min.}

\item{Vin}{Injection flow rate, in L/min.}

\item{temp}{Temperature, degrees Celsius.}

\item{eta}{Dynamic viscosity, in N * s / m^2.}

\item{dims}{A list with the elements \code{L}, \code{b1}, \code{b2}, and \code{z1},
as defined in https://doi.org/10.1016/j.chroma.2018.04.056 (units of cm).}

\item{tol}{Solution is arrived at iteratively; \code{tol} represents the difference between successive iterations.}

\item{maxiter}{Maximum iterations.}
}
\value{
A numeric vector of length one representing the estimated FFF channel thickness.
}
\description{
Estimate the FFF channel thickness using the method described in https://doi.org/10.1021/ac00052a025
}
\examples{
calculate_w(t1 = 21.2, D = .35e-10, Vc = .0015, Vin = 2e-4)
}
