% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_mw.R
\name{calibrate_mw}
\alias{calibrate_mw}
\title{Perform a molecular weight calibration}
\usage{
calibrate_mw(
  time,
  mw,
  type = "quadratic",
  newdata = NULL,
  output = "time",
  predict = TRUE
)
}
\arguments{
\item{time}{A numeric vector of retention times.}

\item{mw}{A numeric vector of molecular weights.}

\item{type}{"linear" or "quadratic"}

\item{newdata}{A vector of retention times or molecular weights.}

\item{output}{"time" or "mw"}

\item{predict}{Logical. Predict from the model?}
}
\value{

}
\description{
Perform a molecular weight calibration
}
