% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fff.R
\name{combine_fff}
\alias{combine_fff}
\title{Combine loaded UV/Vis and ICP-MS data}
\usage{
combine_fff(icp, uv, subtract_blank = TRUE, focus = 10)
}
\arguments{
\item{icp}{A tibble created by \code{load_icp()}}

\item{uv}{A tibble created by \code{load_uv()}}

\item{subtract_blank}{Logical. Subtract a blank from the sample runs?}

\item{focus}{The focusing period, in minutes.}
}
\value{

}
\description{
Combine loaded UV/Vis and ICP-MS data
}
