% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_baseline.R
\name{correct_baseline}
\alias{correct_baseline}
\title{Linear baseline correction with left and right endpoints}
\usage{
correct_baseline(x, left = 10, right = 35, window = 0.2)
}
\arguments{
\item{x}{A data frame.}

\item{left}{The left endpoint.}

\item{right}{The right endpoint.}

\item{window}{The fitting window at each endpoint.}
}
\value{

}
\description{
Linear baseline correction with left and right endpoints
}
