% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_fff.R
\name{deconvolve_fff}
\alias{deconvolve_fff}
\title{Deconvolve incompletely resolved peaks}
\usage{
deconvolve_fff(
  x,
  y,
  h,
  mu,
  s,
  g,
  fn = "skew_gaussian",
  algorithm = "port",
  iter = 300,
  tol = 0.1
)
}
\arguments{
\item{x}{A vector of retention times.}

\item{y}{A vector of detector responses.}

\item{h}{A vector of guesses for peak height, with length equal to the desired number of peaks.}

\item{mu}{A vector of guesses for peak means, with length equal to the desired number of peaks.}

\item{s}{A vector of guesses for peak standard deviations, with length equal to the desired number of peaks.}

\item{g}{A vector of guesses for peak shape parameters, with length equal to the desired number of peaks.}

\item{fn}{Function for peak fitting: skew Gaussian or exponentially modified Gaussian ("skew_gaussian" or "emg").}

\item{algorithm}{Refer to nls() in the stats:: package.}

\item{iter}{Refer to nls() in the stats:: package.}

\item{tol}{Refer to nls() in the stats:: package.}
}
\value{

}
\description{
Deconvolve incompletely resolved peaks
}
