% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_peak.R
\name{integrate_peak}
\alias{integrate_peak}
\title{Integrate FFF peaks}
\usage{
integrate_peak(x, y, injvol = 0.001, flowrate = 0.001)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{injvol}{numeric}

\item{flowrate}{numeric}
}
\value{

}
\description{
Integrate FFF peaks
}
