% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_direct_quant.R
\name{load_direct_quant}
\alias{load_direct_quant}
\title{Read and clean ICP-MS data files generated by the iCAP-RQ}
\usage{
load_direct_quant(path, int_stds = c("45Sc", "115In", "159Tb"))
}
\arguments{
\item{path}{Relative path to the ICP-MS (.xlsx) input files.}

\item{int_stds}{A character vector of internal standards to be removed from the output.}
}
\value{
A tibble with columns \code{date}, \code{sample}, \code{value}, and \code{param}.
}
\description{
Read and clean ICP-MS data files generated by the iCAP-RQ
}
\examples{
path <- system.file("extdata", package = "fffprocessr")
load_direct_quant(path = paste0(path, "/direct_quantification"))
}
