% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_icp.R
\name{load_icp}
\alias{load_icp}
\title{Load FFF-ICP files (with extension .csv)}
\usage{
load_icp(path, calibrate = TRUE)
}
\arguments{
\item{path}{The relative path to the ICP-MS (.csv) files.}

\item{calibrate}{Logical. Convert the raw instantaneous detector counts per second to instantaneous concentrations?}
}
\value{

}
\description{
Load FFF-ICP files (with extension .csv)
}
