% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mals.R
\name{load_mals}
\alias{load_mals}
\title{Load MALS data files (.txt)}
\usage{
load_mals(
  path,
  angles = c(7, 12, 20, 28, 36, 44, 52, 60, 68, 76, 84, 90, 100, 108, 116, 124, 132,
    140, 148, 156, 164)
)
}
\arguments{
\item{path}{The relative path to the MALS (.txt) files.}

\item{angles}{MALS detector angles}
}
\value{

}
\description{
Load MALS data files (.txt)
}
