% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_uv.R
\name{load_uv}
\alias{load_uv}
\title{Title}
\usage{
load_uv(path, nm1 = X2, nm2 = X4, nm3 = X6)
}
\arguments{
\item{path}{The relative path to the ICP-MS (.csv) files.}

\item{nm1}{Name of column 2.}

\item{nm2}{Name of column 4.}

\item{nm3}{Name of column 6.}
}
\value{

}
\description{
Title
}
