% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_maxima.R
\name{peak_maxima}
\alias{peak_maxima}
\title{Detect peak maxima}
\usage{
peak_maxima(data, focus = 10, k = 35, peaks = 1)
}
\arguments{
\item{data}{A tibble returned by \code{load_icp()}, \code{load_uv()}, \code{load_mals()}, or \code{combine_fff()}}

\item{focus}{Focus period, in minutes.}

\item{k}{The basis dimension of the generalized additive model used to detect peak minima.}

\item{peaks}{Number of peaks.}
}
\value{
A tibble of peak retention times and the corresponding maxima in detector response.
}
\description{
Detect peak maxima
}
\examples{
path <- system.file("extdata", package = "fffprocessr")
data <- combine_fff(load_icp(path))
data <- data[data$param == "65Cu", ]
peak_maxima(data, peaks = 3)
}
