% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rg.R
\name{calculate_rg}
\alias{calculate_rg}
\title{Calculate the radius of gyration from FFF-MALS data}
\usage{
calculate_rg(data, window = 0.1, method = "zimm")
}
\arguments{
\item{data}{A data frame.}

\item{window}{The time interval on which to return calculated radii.}

\item{method}{Use method "zimm" or "watt".}
}
\value{

}
\description{
Calculate the radius of gyration from FFF-MALS data
}
