% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rg.R
\name{calculate_rg}
\alias{calculate_rg}
\title{Calculate the radius of gyration from FFF-MALS data}
\usage{
calculate_rg(
  data,
  window = 0.1,
  method = "zimm",
  order = 2,
  calibration = fffprocessr:::mals_calib
)
}
\arguments{
\item{data}{A tibble generated by \code{load_mals()}.}

\item{window}{The time interval on which to return calculated radii.}

\item{method}{Use method "zimm" or "watt". Method "watt" is experimental.}

\item{order}{Order of polynomial fit for method = "watt"}

\item{calibration}{A data frame with three columns: number (angle number), theta (angle), and
coefficient (MALS instrument calibration constant).}
}
\value{
A tibble generated by \code{load_mals()} with additional columns 'timeslice', 'theta',
'rayleigh_ratio', and either 'rg_zimm' or 'rg_watt'.
}
\description{
Calculate the radius of gyration from FFF-MALS data
}
\examples{
path <- system.file("extdata/mals", package = "fffprocessr")
mals_data <- load_mals(path = path)
calculate_rg(mals_data)
}
