% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rh.R
\name{calculate_rh}
\alias{calculate_rh}
\title{Calculate hydrodynamic radii}
\usage{
calculate_rh(
  t1,
  focus = 10,
  transition = 1,
  w = 500/1e+06,
  Vc = 0.0025,
  Vout = 0.001,
  Vin = 5e-04,
  temp = 25,
  eta = 0.00089
)
}
\arguments{
\item{t1}{Retention time, in minutes.}

\item{focus}{Focus period, in minutes.}

\item{transition}{Transition time between focusing and elution, in minutes.}

\item{w}{Channel thickness, in metres.}

\item{Vc}{Cross-flow rate, in L/min.}

\item{Vout}{Detector flow rate, in L/min.}

\item{Vin}{Injection flow rate, in L/min.}

\item{temp}{Temperature, degrees Celsius.}

\item{eta}{Dynamic viscosity, in N * s / m^2.}
}
\value{
A vector of hydrodynamic radii in metres.
}
\description{
Calculate hydrodynamic radii
}
\examples{
tr <- seq(11.6, 15.6, by = .1)
calculate_rh(tr)
}
