% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fff.R
\name{combine_fff}
\alias{combine_fff}
\title{Combine loaded UV-MALS and ICP-MS data}
\usage{
combine_fff(icp = NULL, uv = NULL, subtract_blank = TRUE, focus = 10)
}
\arguments{
\item{icp}{A tibble created by \code{load_icp()}, or}

\item{uv}{A tibble created by \code{load_uv()}, or both.}

\item{subtract_blank}{Logical. Subtract a blank from the sample runs?}

\item{focus}{Focusing period, in minutes.}
}
\value{
A tibble with the columns date', 'sample', 'param', 'time', 'conc', and 'three_sigma'.
The final column, 'three_sigma' represents three times the standard deviation of blank runs
corresponding to each date and analyte.
}
\description{
Combine loaded UV-MALS and ICP-MS data
}
\examples{
path <- system.file("extdata", package = "fffprocessr")
combine_fff(load_icp(path), load_uv(path, UV254_1))
}
