% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_baseline.R
\name{correct_baseline}
\alias{correct_baseline}
\title{Linear baseline correction with left and right endpoints}
\usage{
correct_baseline(
  x,
  left = 10,
  right = 35,
  window = 0.2,
  group_vars = c("date", "sample", "param"),
  x_var = "time",
  y_var = "conc"
)
}
\arguments{
\item{x}{A tibble of the type returned by \code{load_icp()}.}

\item{left}{The left endpoint.}

\item{right}{The right endpoint.}

\item{window}{The fitting window at each endpoint.}

\item{group_vars}{Grouping variables for input to \code{dplyr::group_by()}. Set to NULL when there are no grouping variables.}

\item{x_var}{Name of x variable.}

\item{y_var}{Name of y variable.}
}
\value{
A tibble of the type returned by \code{load_icp()}, with the column 'conc' modified
by a linear baseline correction.
}
\description{
Linear baseline correction with left and right endpoints
}
\examples{
data <- data.frame(
  date = as.Date("2021-01-01"),
  sample = "a",
  param = "56Fe",
  time = seq(0, 30, by = .1),
  conc = runif(301, 0, 1)
)
data$conc <- data$conc + data$time
correct_baseline(data, left = .2, right = 29.8)
}
