% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_fff.R
\name{deconvolve_fff}
\alias{deconvolve_fff}
\title{Deconvolve incompletely resolved peaks}
\usage{
deconvolve_fff(
  x,
  y,
  h,
  mu,
  s,
  g = NULL,
  fn = "skew_gaussian",
  algorithm = "port",
  iter = 300,
  tol = 0.1
)
}
\arguments{
\item{x}{A vector of retention times.}

\item{y}{A vector of detector responses.}

\item{h}{A vector of guesses for peak height, with length equal to the desired number of peaks.}

\item{mu}{A vector of guesses for peak means, with length equal to the desired number of peaks.}

\item{s}{A vector of guesses for peak standard deviations, with length equal to the desired number of peaks.}

\item{g}{A vector of guesses for peak shape parameters, with length equal to the desired number of peaks. Silently
ignored if \verb{fn == "normal}.}

\item{fn}{Function for peak fitting: Gaussian, skew Gaussian or exponentially modified Gaussian ("normal",
"skew_gaussian" or "emg").}

\item{algorithm}{Refer to nls() in the 'stats' package.}

\item{iter}{Refer to nls() in the 'stats' package.}

\item{tol}{Refer to nls() in the 'stats' package.}
}
\value{
A list containing the nls() model, the fitted model evaluated at x, and a tibble with a
column for each of the component peaks evaluated at x.
}
\description{
Deconvolve incompletely resolved peaks
}
\examples{
set.seed(0)
x <- seq(0, 10, length.out = 100)
# two gaussian peaks:
y <- exp(-(x - 3) ^ 2 / 2) + exp(-(x - 7) ^ 2 / 2) + rnorm(100, 0, .1)
deconvolve_fff(x, y, h = c(1, 1), mu = c(3, 7), s = c(1, 1), fn = "normal")
}
