% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_peak.R
\name{integrate_peak}
\alias{integrate_peak}
\title{Integrate FFF peaks}
\usage{
integrate_peak(x, y, injvol = 0.001, flowrate = 0.001)
}
\arguments{
\item{x}{A numeric vector of retention times, in minutes.}

\item{y}{A numeric vector of detector responses, matching x in length.}

\item{injvol}{The injection volume in litres.}

\item{flowrate}{The detector flow rate in litres per minute.}
}
\value{
A number representing the integrated peak area.
}
\description{
Integrate FFF peaks
}
\examples{
x <- seq(0, 10, by = .1)
y <- exp(-(x - 5) ^ 2 / 2) # gaussian peak
integrate_peak(x, y)
}
