% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_icp.R
\name{load_icp}
\alias{load_icp}
\title{Load FFF-ICP data files with extension .csv}
\usage{
load_icp(
  path,
  calibrate = TRUE,
  date_regex = "\\\\d{4}-\\\\d{2}-\\\\d{2}",
  date_format = "\%Y-\%m-\%d",
  calib_path = NULL,
  keywords = NULL,
  data_format = "x-series II"
)
}
\arguments{
\item{path}{Relative path to the ICP-MS (.csv) input files.}

\item{calibrate}{Logical. Convert the raw instantaneous detector counts per second to
instantaneous concentrations? This requires a table for each date, in .xlsx format, with the
following columns: \code{file}, a character vector of filenames; \code{element}, a character vector
describing each calibration curve (e.g., \verb{56Fe}); \code{defined}, a numeric vector of defined
standard concentrations; and \code{mean_cps} a numeric vector of mean counts per second at each
concentration}

\item{date_regex}{An optional regular expression for extracting non-ISO dates from filenames.}

\item{date_format}{An optional non-standard date format corresponding to the output of
\code{date_regex} (see \code{?strptime}).}

\item{calib_path}{Optional. Use if the relative path to the ICP-MS calibration files differs
from the relative path to the data files.}

\item{keywords}{An optional vector of pattern matches to pass to \code{stringr::str_detect()} that tell \code{load_icp()}
which files to load. These can be regular expressions.}

\item{data_format}{Selects an appropriate \code{readr} function based on the data format. Current
options are "x-series II" and "iCAP-RQ".}
}
\value{
A tibble with the columns 'file', 'date', 'param', 'time', and 'conc'.
}
\description{
Load FFF-ICP data files with extension .csv
}
\examples{
path <- system.file("extdata", package = "fffprocessr")
load_icp(path = path)
}
