% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mals.R
\name{load_mals}
\alias{load_mals}
\title{Load FFF-MALS data files with extension .txt}
\usage{
load_mals(
  path,
  angles = fffprocessr:::angles,
  date_regex = "\\\\d{4}-\\\\d{2}-\\\\d{2}",
  date_format = "\%Y-\%m-\%d",
  keywords = NULL,
  angle_names = "ls\\\\d+-\\\\d+",
  ...
)
}
\arguments{
\item{path}{The relative path to the .txt input files.}

\item{angles}{MALS detector angles.}

\item{date_regex}{An optional regular expression for extracting dates from filenames.}

\item{date_format}{An optional non-standard date format corresponding to the output of
\code{date_regex} (see \code{?strptime}).}

\item{keywords}{An optional vector of pattern matches to pass to \code{stringr::str_detect()} that tell \code{load_mals()}
which files to load. These can be regular expressions.}

\item{angle_names}{A regular expression specifying the naming convention for MALS data files. The default is \verb{ls\\\\d+-\\\\d+},
or "ls" followed by the range of angles the file contains (e.g., "ls7-20").}

\item{...}{Other arguments passed on to \code{read_table()}}
}
\value{
A tibble with the columns 'file', 'date', 'sample', 'param', 'time', and 'conc'.
}
\description{
Load FFF-MALS data files with extension .txt
}
\examples{
path <- system.file("extdata/mals", package = "fffprocessr")
load_mals(path = path)
}
