% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_uv.R
\name{load_uv}
\alias{load_uv}
\title{Load FFF-UV/Vis files (.txt)}
\usage{
load_uv(path, nm1, nm2, nm3)
}
\arguments{
\item{path}{The relative path to the UV (.txt) files.}

\item{nm1}{Name of column 2.}

\item{nm2}{Name of column 4.}

\item{nm3}{Name of column 6.}
}
\value{

}
\description{
Load FFF-UV/Vis files (.txt)
}
