% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_uv.R
\name{load_uv}
\alias{load_uv}
\title{Load FFF-UV-MALS data files with extension .txt}
\usage{
load_uv(
  path,
  nm1 = "X2",
  nm2 = "X4",
  nm3 = "X6",
  date_regex = "\\\\d{4}-\\\\d{2}-\\\\d{2}",
  date_format = "\%Y-\%m-\%d",
  keywords = NULL,
  ...
)
}
\arguments{
\item{path}{Relative path to the .txt input files.}

\item{nm1}{Name of column 2 in the input files.}

\item{nm2}{Name of column 4 in the input files.}

\item{nm3}{Name of column 6 in the input files.}

\item{date_regex}{An optional regular expression for extracting dates from filenames.}

\item{date_format}{An optional non-standard date format corresponding to the output of
\code{date_regex} (see \code{?strptime}).}

\item{keywords}{An optional vector of pattern matches to pass to \code{stringr::str_detect()} that tell \code{load_uv()}
which files to load. These can be regular expressions.}

\item{...}{Other arguments passed on to \code{read_table()}}
}
\value{
A tibble with the columns 'file', 'date', 'param', 'time', and 'conc'.
}
\description{
Load FFF-UV-MALS data files with extension .txt
}
\examples{
path <- system.file("extdata", package = "fffprocessr")
load_uv(path = path, UV254_1, UV254_2, LS90)
}
