% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_maxima.R
\name{peak_maxima}
\alias{peak_maxima}
\title{Detect peak maxima}
\usage{
peak_maxima(
  data,
  focus = 10,
  k = 35,
  peaks = 1,
  n = 1,
  method = "gam",
  max_iter = 20,
  group_vars = c("date", "sample", "param"),
  x_var = "time",
  y_var = "conc"
)
}
\arguments{
\item{data}{A tibble returned by \code{load_icp()}, \code{load_uv()}, \code{load_mals()}, or \code{combine_fff()}}

\item{focus}{Peaks occurring at x values less than or equal to this value are not considered. For FFF data, this is the focus period.}

\item{k}{The basis dimension of the generalized additive model used to detect peak minima.}

\item{peaks}{Number of peaks.}

\item{n}{Before detecting peak minima, compute an n-point moving average of the detector response using
\code{stats::filter(x, filter = rep(1/n, n))}.}

\item{method}{Algorithm for peak detection. Either "gam", which separates peaks by the minima between
them, or "sigma" which recursively calculates the standard deviation by eliminating values greater
than three standard deviations. N.B., method = "sigma" silently removes NAs from x and y.}

\item{max_iter}{Maximum iterations for method = "sigma".}

\item{group_vars}{Grouping variables for input to \code{dplyr::group_by()}.}

\item{x_var}{Name of x variable.}

\item{y_var}{Name of y variable.}
}
\value{
A tibble of peak retention times and the corresponding maxima in detector response.
}
\description{
Detect peak maxima
}
\examples{
path <- system.file("extdata", package = "fffprocessr")
data <- combine_fff(load_icp(path))
data <- data[data$param == "65Cu", ]
peak_maxima(data, peaks = 3)
}
